requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.2.0
beartype==0.17.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
fast-bitrix24==1.5.16
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gspread==6.0.2
httplib2==0.22.0
icontract==2.6.6
idna==3.6
more-itertools==10.2.0
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==65.5.1
six==1.16.0
StrEnum==0.4.15
tqdm==4.66.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
wheel==0.38.4
yarl==1.9.4
